% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_config.R
\name{hail_config}
\alias{hail_config}
\title{Read Hail Configuration}
\usage{
hail_config(config = sparklyr::spark_config())
}
\arguments{
\item{config}{A spark configuration.}
}
\description{
Set configuration for Hail using \code{spark_config()}.
}
