% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_context.R
\name{hail_context}
\alias{hail_context}
\title{Create Hail Context}
\usage{
hail_context(sc)
}
\arguments{
\item{sc}{Spark connection.}
}
\value{
hailContext
}
\description{
Import and initialize Hail using a spark connection.
}
\examples{
library(sparklyr)

sc <- spark_connect(master = "spark://HOST:PORT", config = hail_config())
connection_is_open(sc)
hail_context(sc)
spark_disconnect(sc)

}
