% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_entries.R
\name{hail_entries}
\alias{hail_entries}
\title{Get Entries Field}
\usage{
hail_entries(df)
}
\arguments{
\item{df}{A hail dataframe.}
}
\value{
A spark dataframe.
}
\description{
This function retrieves the entries fields from a hail dataframe and explodes the
columns call, dp and gq.
}
\examples{
\dontrun{
library(sparklyr)

sc <- spark_connect(master = "local", version = "2.4", config = hail_config())

hail_context(sc) \%>\% 
  hail_read_matrix(system.file("extdata/1kg.mt", package = "sparkhail")) \%>\% 
  hail_dataframe() \%>\% 
  hail_entries()
}

}
