% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_get_1kg.R
\name{hail_get_1kg}
\alias{hail_get_1kg}
\title{Download the Dataset Examples}
\usage{
hail_get_1kg()
}
\description{
This function creates an extdata folder and downloads the datasets necessary to
run the examples: 1kg MatrixTable folder and annotations.txt.
}
\examples{
\dontrun{
hail_get_1kg()
}
}
