% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_ids.R
\name{hail_ids}
\alias{hail_ids}
\title{Get Sample Ids}
\usage{
hail_ids(mt)
}
\arguments{
\item{mt}{A MatrixTable object.}
}
\description{
Get the ids from s col key in a MatrixTable.
}
\examples{
\dontrun{
library(sparklyr)


hl <- hail_context(sc)
mt <- hail_read_matrix(hl, system.file("extdata/1kg.mt", package = "sparkhail"))

hail_ids(mt)
}

}
