% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hail_install.R
\name{hail_install}
\alias{hail_install}
\alias{hail_uninstall}
\title{Install Hail and Datasets}
\usage{
hail_install(datasets_examples = TRUE)

hail_uninstall()
}
\arguments{
\item{datasets_examples}{If TRUE, hail will be downloaded along with the datasets to
run the examples. Use FALSE if you just want to install hail.}
}
\description{
Install hail dependencies and datasets to run the examples in documentation.
To remove hail use \code{hail_uninstall}.
}
